+++ tests/unit/data/ar/Order.php
@@ -7,7 +7,7 @@ namespace yiiunit\data\ar;
  *
  * @property integer $id
  * @property integer $customer_id
- * @property integer $create_time
+ * @property integer $created_at
  * @property string $total
  */
 class Order extends ActiveRecord
@@ -68,7 +68,7 @@ class Order extends ActiveRecord
 	public function beforeSave($insert)
 	{
 		if (parent::beforeSave($insert)) {
-			$this->create_time = time();
+			$this->created_at = time();
 			return true;
 		} else {
 			return false;