+++ extensions/codeception/CHANGELOG.md
@@ -1,6 +1,11 @@
 Yii Framework 2 Codeception extension Change Log
 ================================================
 
+- Enh: yii\codeception\TestCase now supports loading and using fixtures via Yii fixture framework (qiangxue)
+- New #1956: Implemented test fixture framework (qiangxue)
+- New: Added yii\codeception\DbTestCase (qiangxue)
+
+
 2.0.0 beta under development
 ----------------------------